fix6

+++ deps/faiss/faiss/gpu/perf/IndexWrapper.h
@@ -1,5 +1,5 @@
-/*
- * Copyright (c) Meta Platforms, Inc. and affiliates.
+/**
+ * Copyright (c) Facebook, Inc. and its affiliates.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
@@ -32,7 +32,7 @@ struct IndexWrapper {
     faiss::Index* getIndex();
 
     void runOnIndices(std::function<void(GpuIndex*)> f);
-    void setNumProbes(size_t nprobe);
+    void setNumProbes(int nprobe);
 };
 
 } // namespace gpu
